+++ application/modules/editor/src/Task/Import/FileParser/Xlf/Namespaces/Tmgr.php
@@ -9,13 +9,13 @@ START LICENSE AND COPYRIGHT
  Contact:  http://www.MittagQI.com/  /  service (ATT) MittagQI.com
 
  This file may be used under the terms of the GNU AFFERO GENERAL PUBLIC LICENSE version 3
- as published by the Free Software Foundation and appearing in the file agpl3-license.txt 
- included in the packaging of this file.  Please review the following information 
+ as published by the Free Software Foundation and appearing in the file agpl3-license.txt
+ included in the packaging of this file.  Please review the following information
  to ensure the GNU AFFERO GENERAL PUBLIC LICENSE version 3 requirements will be met:
  http://www.gnu.org/licenses/agpl.html
   
  There is a plugin exception available for use with this release of translate5 for
- translate5: Please see http://www.translate5.net/plugin-exception.txt or 
+ translate5: Please see http://www.translate5.net/plugin-exception.txt or
  plugin-exception.txt in the root folder of translate5.
   
  @copyright  Marc Mittag, MittagQI - Quality Informatics
@@ -26,17 +26,16 @@ START LICENSE AND COPYRIGHT
 END LICENSE AND COPYRIGHT
 */
 
-/** #@+
- * @author Marc Mittag
- * @package editor
- * @version 1.0
- */
+namespace MittagQI\Translate5\Task\Import\FileParser\Xlf\Namespaces;
+
+use editor_Models_Import_FileParser_SegmentAttributes as SegmentAttributes;
 
 
 /**
  * XLF Fileparser Add On to parse IBM XLF specific stuff
  */
-class editor_Models_Import_FileParser_Xlf_Namespaces_Tmgr extends editor_Models_Import_FileParser_Xlf_Namespaces_Abstract{
+class Tmgr extends AbstractNamespace
+{
     const IBM_XLIFF_NAMESPACE = 'xmlns:tmgr="http://www.ibm.com"';
 
     protected static function isApplicable(string $xliff): bool
@@ -46,18 +45,17 @@ class editor_Models_Import_FileParser_Xlf_Namespaces_Tmgr extends editor_Models_
 
     /**
      * {@inheritDoc}
-     * @see editor_Models_Import_FileParser_Xlf_Namespaces_Abstract::transunitAttributes()
+     * @see AbstractNamespace::transunitAttributes()
      */
-    public function transunitAttributes(array $attributes, editor_Models_Import_FileParser_SegmentAttributes $segmentAttributes): void
+    public function transunitAttributes(array $attributes, SegmentAttributes $segmentAttributes): void
     {
         //FIXME add match rate infos into our matchRateType field!
-        settype($attributes['tmgr:matchratio'], 'integer');
-        $segmentAttributes->matchRate = $attributes['tmgr:matchratio'];
+        $segmentAttributes->matchRate = (int) $attributes['tmgr:matchratio'];
     }
-    
+
     /**
      * {@inheritDoc}
-     * @see editor_Models_Import_FileParser_Xlf_Namespaces_Abstract::useTagContentOnly()
+     * @see AbstractNamespace::useTagContentOnly()
      */
     public function useTagContentOnly(): ?bool
     {